requirements.txt
discord.py==2.5.2
flask==3.1.0
waitress==3.0.2
python-dotenv==1.0.1
requests==2.32.3
jsonschema==4.23.0
psutil==6.1.1
websockets==15.0.1
